Extracted the signup form
